添加向量的加法、减法和数量积、点积等基本运算，自定义错误

+++ src/macros.rs
@@ -4,9 +4,9 @@ macro_rules! array {
         $crate::nd_array::Array::new()
     );
     ($elem:expr; $n:expr) => (
-        $crate::Array::from([$elem; $n])
+        $crate::Array::from([$elem as f64; $n])
     );
     ($($x:expr),+ $(,)?) => (
-        $crate::Array::from(vec![$($x),+])
+        $crate::Array::from(vec![$($x as f64),+])
     );
 }
